Fix #2033
This commit aims to fix the issue described in #2033 by introducing a check into InWorldProcessing#decrementProcessingTime that forces a time recalculation if the time value is less than 0. Time should not be less than 0 in any case that this method is called, as -1 is reserved to indicate that the item does not fit the process type.