Initial implementation of Xaero train map
This implementation follows the same style as the existing Journeymap integration.
There are a couple bugs with this that will hopefully be fixed soon.
Rendering order - The train map renders over all UI elements which are part of the map. This is due to the mixin targeting the tail of render(). However, targeting earlier in the method causes visual errors as the rendering of map sections seems to be multithreaded.
Trains disappearing - When another menu is accessed such as the map settings, the trains on the map seem to disappear for a while. Updating them in the world seems to make them appear again.