Train Map Viewing in other dimensions
Xaero's map allows you to view maps of other dimensions the player is not currently in, so this fixes the wrong train map being visible when viewing maps from other dimensions. When ticking, we check if Xaero's map is open before ticking the rendered dimension as the rendered dimension is never unset outside of Xaero's render function. This causes issues if another map mod is loaded and tries to view the train map as it would tick whatever map Xaero's was last rendering.