Better map open check
Xaero's map can layer multiple transparent screens so checking if the map is open isn't best performed by checking if the current screen is GuiMap. This fixes the map not updating if another menu (such as map settings) is open over the map. We still check if it's GuiMap when checking for mouse clicks on the toggle widget, as if another screen is rendering over it, there may be a button conflict.