Pause Screen Injection
Xaero's world map pauses the game in singleplayer when open which prevents the train map from getting the train positions properly. This adds a mixin which overrides this behaviour by injecting into the head of isPauseScreen(). This allows the map to work perfectly in singleplayer.
Also the mixin for the map now has require = 0 incase of future changes which break the injection.