Accessor, try-catches and fixed map open check
Switched to using an accessor instead of shadow fields, as this apparently is less prone to failure if Xaero's changes field names/types, as the calls to the accessor are wrapped in a try-catch.
Put try-catch blocks around any reference to Xaero's code incase of changes. Currently if an exception is thrown an error log is printed once.
Improved the IsMapOpen() function to check if the parent screen is actually xaero's map as I didn't realize you can access the map settings screen for example without the map being open. The game pausing override injection uses this logic now also for consistency.