Fix unloaded postbox voiding packages
There seems to be an area of chunks where `Level.isLoaded` returns
`false`, but `PostboxBlockEntity.onChunkUnloaded` has not yet been
called. Any trains delivering to that area will use the offline buffer,
but when the block entity is eventually unloaded, it will overwrite the
delivered packages with an older version of the inventory, causing them
to be lost.

I made the following changes to fix this:

- Sync the offline buffer whenever the inventory changes, rather than
  in `onChunkUnloaded`.
- Fix `StationBlockEntity.attachPackagePort` incorrectly replacing
  `GlobalPackagePort` instances with new ones, resetting their data.
- During mail transfer, get the `Level` from `ServerLifecycleHooks`
  rather than `CarriageContraptionEntity`: this ensures a loaded
  postbox will always be discovered, regardless of whether the train
  itself is loaded.

Fixes #7491