Fix #8508 Validate item in ChainPackageInteractionPacket
Adds a check to the ChainPackageInteractionPacket to check if the players main hand matches what was sent to the server.

This not only fixes the dupe but also fixes a packet exploit that would allow malicious actors to spawn in any item by simply providing bogus data when sending the packet to the server.