Fix potion-based fluid recipes.
Creators-of-Create/Create@58876e2 migrated from a custom ingredient
handler for fluids to NeoForge's `SizedFluidIngredient`.

In some places, references were updated to use a `SizedFluidIngredient`
containing a `DataComponentFluidIngredient` in order to preserve
behaviors which would require the presence of specific components
in fluid ingredients (the previous behavior).

In other places, references were updated to use
`SizedFluidIngredient.of(FluidStack stack)`, which discards
components when creating the fluid ingredient.

This resulted in Creators-of-Create/Create#9474, where potion recipes
were effectively created such that a non-fluid potion ingredient
could be mixed with _any_ potion base to create (one of) the possible
outputs when brewing a potion with said ingredient.

This change updates the potion-mixing recipes logic to use a
`DataComponentFluidIngredient` as the base fluid ingredient
inside the `SizedFluidIngredient`, fixing
Creators-of-Create/Create#9474.

This change also updates JEI integration code to similarly use
`DataComponentFluidIngredients`, fixing visualization issues with
potion-bottle and potion-application (e.g., redstone from cinder flour)
recipes.

Adds tests to verify potion brewing and potion-application recipes
function as intended.